requirements.txt
attrs==22.1.0
click==8.1.3
factory-boy==3.2.1
Faker==15.0.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
Werkzeug==2.2.2